Add default parameter type, adjust usages

+++ src/knx/qknxnetipknxaddressesdib.h
@@ -42,7 +42,7 @@ public:
         const QKnxNetIpPayload &load = payload();
 
         T addresses(load.size() / 2, Qt::Uninitialized);
-        for (int i = 0; i < load.size(); i += 2)
+        for (quint16 i = 0; i < load.size(); i += 2)
             addresses.push_back({ QKnxAddress::Type::Individual, load.bytes<QByteArray>(i, 2) });
         return addresses;
